Created Matrix Manager Files and cleaned up
